LPD-66872 use ctService
