+++ modules/apps/change-tracking/change-tracking-service/build.gradle
@@ -10,6 +10,7 @@ dependencies {
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "default"
 	compileOnly group: "jakarta.servlet", name: "jakarta.servlet-api", version: "6.0.0"
+	compileOnly group: "net.htmlparser.jericho", name: "jericho-html", version: "3.4"
 	compileOnly group: "org.osgi", name: "org.osgi.annotation.versioning", version: "1.1.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.cm", version: "1.6.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.4.0"
@@ -18,6 +19,7 @@ dependencies {
 	compileOnly project(":apps:change-tracking:change-tracking-spi")
 	compileOnly project(":apps:change-tracking:change-tracking-store-api")
 	compileOnly project(":apps:dispatch:dispatch-api")
+	compileOnly project(":apps:dynamic-data-mapping:dynamic-data-mapping-api")
 	compileOnly project(":apps:journal:journal-api")
 	compileOnly project(":apps:json-storage:json-storage-api")
 	compileOnly project(":apps:portal-background-task:portal-background-task-api")
